+++ plugins/SebLogiWheel/src/SebLogiWheel/Plugin.cs
@@ -9,6 +9,6 @@ namespace SebLogiWheel
     {
         public const string PluginGuid = "shibe.easydeliveryco.seblogiwheel";
         public const string PluginName = "SebLogiWheel";
-        public const string PluginVersion = "1.0.0";
+        public const string PluginVersion = "1.0.1";
     }
 }